Conflicts:
	src/scip/misc.h
	src/scip/scip.c
	src/scip/scip.h

+++ src/scip/type_misc.h
@@ -163,7 +163,7 @@ typedef struct SCIP_Uf SCIP_UF;
 #define SCIP_DECL_HASHKEYEQ(x) SCIP_Bool x (void* userptr, void* key1, void* key2)
 
 /** returns the hash value of the key */
-#define SCIP_DECL_HASHKEYVAL(x) unsigned int x (void* userptr, void* key)
+#define SCIP_DECL_HASHKEYVAL(x) uint64_t x (void* userptr, void* key)
 
 
 #ifdef __cplusplus
